+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Next version
+
+### 🚀 New
+
+* Support choice parameter type.
+
+
 ## 0.1.0b4 - Mar 10, 2023
 
 ### 🚀 New